Update download-GA-pack.sh
Changes focus on refining the way the script accesses and uses the JSON credentials, with an emphasis on security and efficiency.

Previously, the value for keyJsonFile was derived from the first positional argument ($1) which was written for a file in the workflow.
In the new code, it is derived from the environment variable MIGRATION_JSON_ENCODED. After fetching the data from the variable, it gets base64-decoded to retrieve the JSON contents. This is a significant change as the way credentials are passed to the script has been altered for added security.