Modify extractClaims to work with regex
Modify the `extractClaims` method to select the div tags that contains the `claim-list` class. With the earlier implementation, if someone add a newer class to the div tag, the tests will get failed. With this implementation, it will check a regex match to identify the claim list, instead of string comparison.